+++ tests/flytekit/unit/core/test_async_more_semantics.py
@@ -2,11 +2,10 @@ import asyncio
 import typing
 import pytest
 
-from flytekit import task
+from flytekit.core.task import task, eager
 from flytekit.configuration import Config
 from flytekit.core.context_manager import FlyteContextManager
 from flytekit.core.data_persistence import FileAccessProvider
-from flytekit.experimental.eager_function import eager
 from flytekit.remote.remote import FlyteRemote
 from flytekit.utils.asyn import loop_manager
