+++ tests/flytekit/unit/core/test_async_with_wflp.py
@@ -2,13 +2,12 @@ import asyncio
 import typing
 import pytest
 
-from flytekit.core.task import task
+from flytekit.core.task import task, eager
 from flytekit.core.workflow import workflow
 from flytekit.configuration import Config
 from flytekit.core.context_manager import FlyteContextManager
 from flytekit.core.launch_plan import LaunchPlan
 from flytekit.core.data_persistence import FileAccessProvider
-from flytekit.experimental.eager_function import eager
 from flytekit.remote.remote import FlyteRemote
 from flytekit.utils.asyn import loop_manager
